% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{analyzeCorrelation}
\alias{analyzeCorrelation}
\title{Title}
\usage{
analyzeCorrelation(
  data_frame,
  first_column_name,
  second_column_name,
  boxcox_transform = FALSE
)
}
\arguments{
\item{data_frame}{}

\item{first_column_name}{}

\item{second_column_name}{}

\item{boxcox}{}
}
\value{

}
\description{
Title
}
