% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorVarCorrelation.R
\name{factorVarCorrelation}
\alias{factorVarCorrelation}
\title{Title}
\usage{
factorVarCorrelation(data_frame, first_column_name, second_column_name)
}
\arguments{
\item{data_frame}{}

\item{first_column_name}{}

\item{second_column_name}{}
}
\value{

}
\description{
Title
}
