% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{numbins}
\alias{numbins}
\title{Generates the number of bins to use for a histogram}
\usage{
numbins(x)
}
\arguments{
\item{x}{A vector of data points}
}
\value{

}
\description{
Generates the number of bins to use for a histogram
}
