% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleEDA.R
\name{simpleEDA}
\alias{simpleEDA}
\title{A function that performs an EDA for a regression model over columns of a 
data frame}
\usage{
simpleEDA(data_frame, response_column_name, predictor_column_names)
}
\arguments{
\item{data_frame}{The data frame to analyze}

\item{response_column_name}{The name of the response variable column}

\item{predictor_column_names}{The names of all predictor variable columns}
}
\value{

}
\description{
A function that performs an EDA for a regression model over columns of a 
data frame
}
