% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{analyzeCorrelation}
\alias{analyzeCorrelation}
\title{A function that analyzes the correlation between two columns of a data frame}
\usage{
analyzeCorrelation(
  data_frame,
  first_column_name,
  second_column_name,
  boxcox_transform = FALSE
)
}
\arguments{
\item{data_frame}{The data frame to analyze}

\item{first_column_name}{The name of the first column to compare}

\item{second_column_name}{The name of the second column to compare}

\item{boxcox_transform}{Whether or not to perform the Box-Cox transform}
}
\value{

}
\description{
A function that analyzes the correlation between two columns of a data frame
}
