% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorVarQuantVarCorrelation.R
\name{factorVarQuantVarCorrelation}
\alias{factorVarQuantVarCorrelation}
\title{Title}
\usage{
factorVarQuantVarCorrelation(data_frame, factor_column_name, quant_column_name)
}
\arguments{
\item{data_frame}{}

\item{factor_column_name}{}

\item{quant_column_name}{}
}
\value{

}
\description{
Title
}
