% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantVarCorrelation.R
\name{quantVarCorrelation}
\alias{quantVarCorrelation}
\title{Title}
\usage{
quantVarCorrelation(
  data_frame,
  first_column_name,
  second_column_name,
  boxcox_transform = FALSE
)
}
\arguments{
\item{data_frame}{}

\item{first_column_name}{}

\item{second_column_name}{}

\item{boxcox}{- A TRUE/FALSE value indicating whether or not to perform a boxcox transformation
and generate a plot}
}
\value{

}
\description{
Title
}
