% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleEDA.R
\name{simpleEDA}
\alias{simpleEDA}
\title{Title}
\usage{
simpleEDA(data_frame, response_column_name, predictor_column_names)
}
\arguments{
\item{data_frame}{}

\item{response_column_name}{}

\item{predictor_column_names}{}
}
\value{

}
\description{
Title
}
